refactor: layer 1 by app-infra-interfaces-shared
