+++ src/app.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { DatabaseModule } from './database/database.module';
+import { DatabaseModule } from 'src/interfaces/modules/database.module';
 
 @Module({
   imports: [DatabaseModule],
