+++ src/interfaces/entities/user.entity.ts
